Fixes error when searching for *.runtimeconfig.json (#84)
* Fixes error when searching for *.runtimeconfig.json

- During .Net Execute the *.runtimeconfig.json must be present in the
working directory, this will be true is the buildpack built the app
because all build artifacts are put into the working directory. If a
user were to supply a pre-published app, that app must have the
*.runtimeconfig.json in the top level of the app directory.

* Fixes error when searching for *.runtimeconfig.json

- During .Net Execute the *.runtimeconfig.json must be present in the
working directory, this will be true is the buildpack built the app
because all build artifacts are put into the working directory. If a
user were to supply a pre-published app, that app must have the
*.runtimeconfig.json in the top level of the app directory.

* Fixes start command